pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aria"
version = "0.1.0"
description = "Autonomous Research Intelligence Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "openai>=1.60.0",
    "anthropic>=0.40.0",
    "pyyaml>=6.0.2",
    "yfinance>=0.2.50",
    "pandas>=2.2.0",
    "ddgs>=9.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0"]

[project.scripts]
aria = "aria.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["aria*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
